Adding some doc/sample recipe Moving config into separate file
