+++ metadata-ingestion/docs/sources/unity-catalog/unity-catalog.md
@@ -1,0 +1,11 @@
+### Setup
+
+1. Generate a Databrick Personal Access token following the guide here: https://docs.databricks.com/dev-tools/api/latest/authentication.html#generate-a-personal-access-token
+2. Get your workspace Id where Unity Catalog is following: https://docs.databricks.com/workspace/workspace-details.html#workspace-instance-names-urls-and-ids
+3. Check our sample recipe and replace Token and Workspace Id with the ones above.
+
+## Troubleshooting
+
+`### [Common Issue]
+
+[Provide description of common issues with this integration and steps to resolve]
